requirements.txt
anthropic>=0.40.0
google-genai>=1.0.0
pypdf>=4.0.0
pyyaml>=6.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.7
